+++ cgogn/modeling/examples/curve.cpp
@@ -113,8 +113,8 @@ void Viewer::keyPressEvent(QKeyEvent *ev)
 			v0 = cgogn::modeling::generate_curve<Vec3>(map_,
 											 vertex_position_,
 											 200,
-											 0.,
-											 10.*M_PI,
+											 0,
+											 float32(10*M_PI),
 											 [&](float32 t){return Vec3(t*std::cos(t),t*std::sin(t),t); });
 		}
 			break;
@@ -126,8 +126,8 @@ void Viewer::keyPressEvent(QKeyEvent *ev)
 			v0 = cgogn::modeling::generate_curve<Vec3>(map_,
 											 vertex_position_,
 											 200,
-											 0.,
-											 4.*M_PI,
+											 0,
+											 float32(4*M_PI),
 											 [&](float32 t){return Vec3(std::sin(2*t),std::cos(t),t); });
 		}
 			break;
@@ -139,8 +139,8 @@ void Viewer::keyPressEvent(QKeyEvent *ev)
 			v0 = cgogn::modeling::generate_curve<Vec3>(map_,
 											 vertex_position_,
 											 200,
-											 0.,
-											 10.*M_PI,
+											 0,
+											 float32(10*M_PI),
 											 [&](float32 t){return Vec3(std::cos(t),std::sin(t),t); });
 		}
 			break;
@@ -165,8 +165,8 @@ void Viewer::keyPressEvent(QKeyEvent *ev)
 			v0 = cgogn::modeling::generate_curve<Vec3>(map_,
 											 vertex_position_,
 											 200,
-											 0.,
-											 10.*M_PI,
+											 0,
+											 float32(10*M_PI),
 											 [&](float32 t){return Vec3(t,t*std::cos(t),t*std::sin(t)); });
 		}
 			break;
@@ -178,8 +178,8 @@ void Viewer::keyPressEvent(QKeyEvent *ev)
 			v0 = cgogn::modeling::generate_curve<Vec3>(map_,
 											 vertex_position_,
 											 200,
-											 0.,
-											 2.*M_PI,
+											 0,
+											 float32(2*M_PI),
 											 [&](float32 t){return Vec3(3.*std::cos(t)+std::cos(10.*t)*std::cos(t),
 																		3.*std::sin(t)+std::cos(10.*t)*std::sin(t),
 																		std::sin(10.*t)); });
@@ -193,8 +193,8 @@ void Viewer::keyPressEvent(QKeyEvent *ev)
 			v0 = cgogn::modeling::generate_curve<Vec3>(map_,
 											 vertex_position_,
 											 200,
-											 0.,
-											 2.*M_PI,
+											 0,
+											 float32(2*M_PI),
 											 [&](float32 t){return Vec3(t*std::cos(t), t, t*std::cos(t)); });
 		}
 			break;
@@ -206,8 +206,8 @@ void Viewer::keyPressEvent(QKeyEvent *ev)
 			v0 = cgogn::modeling::generate_curve<Vec3>(map_,
 											 vertex_position_,
 											 200,
-											 0.,
-											 2.*M_PI,
+											 0,
+											 float32(2*M_PI),
 											 [&](float32 t){return Vec3(5.*std::cos(t)-std::cos(5.*t), 5.*std::sin(t)-std::sin(5.*t), t); });
 		}
 			break;
@@ -219,9 +219,9 @@ void Viewer::keyPressEvent(QKeyEvent *ev)
 			v0 = cgogn::modeling::generate_curve<Vec3>(map_,
 											 vertex_position_,
 											 200,
-											 0.,
-											 2.*M_PI,
-											 [&](float32 t){return Vec3(2.*std::sin(3.*t)*std::cos(t), 2.*std::sin(3.*t)*std::sin(t), 0.); });
+											 0,
+											 float32(2*M_PI),
+											 [&](float32 t){return Vec3(2.*std::sin(3.*t)*std::cos(t), 2.*std::sin(3.*t)*std::sin(t), 0); });
 		}
 			break;
 		case Qt::Key_9:
@@ -232,8 +232,8 @@ void Viewer::keyPressEvent(QKeyEvent *ev)
 			v0 = cgogn::modeling::generate_curve<Vec3>(map_,
 											 vertex_position_,
 											 200,
-											 0.,
-											 2.*M_PI,
+											 0,
+											 float32(2*M_PI),
 											 [&](float32 t){return Vec3(2.*std::sin(3.*t)*std::cos(t), 2.*std::sin(3.*t)*std::sin(t), std::sin(3.*t)); });
 		}
 			break;
@@ -251,7 +251,7 @@ void Viewer::keyPressEvent(QKeyEvent *ev)
 		drawer_->vertex3fv(vertex_position_[v]);
 	});
 	drawer_->end();
-	drawer_->line_width_aa(0.8);
+	drawer_->line_width_aa(0.8f);
 	drawer_->begin(GL_LINES);
 	drawer_->color3f(0.0,0.0,0.0);
 	map_.foreach_cell([&](UndirectedGraph::Edge e)
